Added in relative positions as input representations
Added the choice between relative positions as inputs and absolute
positions. In the case of relative positions objects are represented as
3 inputs: a distance input, which is simply the sq distance from avatar
to object, and 2 direction inputs. These takes the values 1,0 and -1.
These two together code the most important direction of the object. So
if an object is really close in the x direction then the x direction
input will show the direction and the y direction will be 0. The idea is
to allow the agent to easily learn "i want to get closer to this object"
or "I want to get away from this object". This allows it to for example
learn staying away from an object easily by learning "if sqdist<k then
go in the opposite direction of x or y". This basically does the
preprocessing of going from the avatar location and object location as x
and y coordinates to distance and direction for the agent, it would
likely have to learn that itself otherwise.