Added many fitness function methods and added different object representation
All these new different fitness functions are the ones we will compare
to eachother. The different object representation is to do
distance+xdirection+ydirection instead of distance+absolute x and
absolute y coordinate. This should be a lot easier to learn from, and my
tests show this indeed improves performance.